🐞 - `valid_release.tgz` complies with new validations
The pr https://github.com/cloudfoundry/bosh/pull/2479 introduced a
validation which asserts that the `name` key in `job.MF` must match the
Job's name. The test fixture `valid_release.tgz` did not have an `name`
key it any of the `job.MF` files.
This commit updates the contents of `valid_release.tgz` so that the
`name` key is present in `job.MF`. In addition to the required elements,
the `.tgz` contians the un-archived job directories with the contents of
the various $JOB.tgz files which are included in the release.

Without these fixes on is likely to see the following error when running
the integration suite:
```
Task 6 | 04:43:18 | Creating new jobs: cacher/1a (00:00:00)
                 L Error: Inconsistent name for job 'cacher'(exptected: 'cacher', got: '')
Task 6 | 04:43:18 | Error: Inconsistent name for job 'cacher'(exptected: 'cacher', got: '')
```

See
https://github.com/cloudfoundry/bosh/pull/2479#issuecomment-1888218237
for more context.