Remove brat expectation checking for blobstore creds at top level of settings.json
The docker-cpi was udpated to the latest version of bosh-cpi-go.

https://github.com/cloudfoundry/bosh-cpi-go/commit/00bd87aba0c486661e6a5ae08efb250923f7de95 was
done to remove the blobstore section from the top level of the cpi settings in favor of using
the blobstore settings nested in the env that are provided directly by the bosh director.

This updates the test to no longer look for the top level blobstore settings.

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>