Increase agent timeout for `start` command
We found a customer whose Windows vms were taking a long time to start
all the jobs.  By default, the agent client waits 45 seconds for a
response.  This default has not changed for 10 years or so, when it was
bumped from 30 -> 45 seconds in 3d895cd.

Most notably, the `start` action is synchronous, so it has to complete
fully before returning to the director.

Within the Windows lifecycle, there is a wrapping process that uses
WinSW to encapsulate bosh jobs as Windows services.  In our test
environments, we had each "wrapping" job taking 1-2 seconds, so the TASW
tile (with ~10 jobs) took 25 seconds or so.  The customer has TKGi with
a similar amount of jobs, but each took 4-5 seconds, which ended up
exceeding the default 45 second timeout.