Remove EventMachine from DNS sync scripts
The DNS sync scripts were setting up EventMachine, but seemingly never
made use of it.

Additionally, the AgentBroadcaster (which is used both inside the DNS
sync scripts and in the Director during deploy) had a reference to a
"EmReactorLoop" class. It appears this class used to use EventMachine,
but now simply yields whatever block is passed to it. Since it's just
calling a block method, it has been removed in favor of just the code in
the block.