Update release tar specs to ignore macOS resource-fork files
Files with the prefix `._` are created in TAR files to represent macOS
resource forks. This update modifies a helper method which expands
release tar files so that it excludes these files from the list used in
specs for comparison. It also updates a spec to use this helper method
rather than a custom tar extraction invocation.