Ensure Health Monitor HTTP connections are closed
When EventMachine was removed from Health Monitor in favor of
async-http, it neglected to explicitly close the Async::Http::Client
objects. This results in connections to the Director on port 25555 to
constantly accumulate and remain in a CLOSE_WAIT state. This can result
in problems with NATS, such as the bosh-nats-sync process being unable
to connect to the Director to get a list of valid clients.