🐞 integration tests use modern UAA
Previous to this commit, the integration tests used a very very old
version of UAA (3.5.0) from an s3 bucket that needed Java 8.  Updates to
the integration Docker image that raised the JDK from 8->17 caused this
very old version to explode.

This commit represents a major refactoring of how UAA runs in the
integration specs.
* The UAA version now tracks the version in bosh-deployment
* There is only 1 UAA started, as opposed to 1 per spec process
* The UAA version is from uaa-release, and is templated accordingly

Java upgrade PRs:
* #2514
* #2515
* #2516

Co-authored-by: Aram Price <aram.price@broadcom.com>
Co-authored-by: Chris Selzo <chris.selzo@broadcom.com>