Stop depending on bosh-dependencies S3 bucket
Previously, we were downloading very old, pre-compiled artifacts from a
bosh-dependencies bucket in S3 as part of integration tests. The only
remaining test was checking that an unspecified old version of the agent
responded with an error to a newer NATS message.

This also removes a script to install an old version of Tomcat that was
no longer being used.