Health Monitor continues to log after errors
When health_monitor is sending metrics to, for example, graphite, and
there is a network error (e.g. `Errno::EPIPE`), it stops sending metrics
until health_monitor is restarted.

This was a regression due to replacing the `EventMachine` gem with the `Async`
gem.

This commit fixes the regression by, when a network error occurs,
attempting to re-establish the connection and continue sending data.

The attempt to re-establish the connection follows the same retry &
backoff logic as when establishing the initial connection.

Note: we feel the method `unbind` is poorly named; it should be named
`close_old_and_open_new_connection`.

[fixes #2522]

[#187636407]

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>