🐞 Health Monitor properly feeds plugins
Health Monitor needs the gem "async/io" to properly feed messages to
plugins such as Graphite.

Previously "async/io" was a dependency of "async/http", so it did not
need to be explicitly required; however, that has recently changed.

The symptom we witnessed was health messages were NOT forwarded to
Graphite using the latest Director. The error was tracked to the
following in `tcp_connection.rb`:

```
uninitialized constant Async::IO
```

This commit fixes that error by explicitly requiring "async/io"

Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>
Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>
Signed-off-by: Ming Xiao <ming.xiao@broadcom.com>
Signed-off-by: Brian Cunnie <brian.cunnie@broadcom.com>