Follow up to [#2533]
The tombstone creation happens in a larger transaction, so if you both insert and clean tombstone records at the
same time you can end up in a deadlock state if that is happening in multiple threads at the same time.

This commit moves the tombstone cleanup into a scheduled job that runs once a day instead.