Brings up Health Monitor HTTP server faster
This commit adds a guard around the '/unresponsive_agents' endpoint so
that it will return a "not successful" HTTP status code (in this case,
503) if the initial "query all Deployments and their Instances" action
run by the Director has not yet completed.

To support that, it pushes down the guts of the fetch_deployments
function into the InstanceManager class.

This commit also moves the start of the Health Monitor HTTP server to
nearly the top of the 'Monitor#run' function. This should get the
'/healthz' endpoint started as quickly as possible so that we don't get
terminated by monit just because querying the state of the
Director-managed deployments takes longer than ten seconds.

Why do this?

We've discovered that Monit gives a monitored service ~10 seconds to
bring up its health-checking HTTP server before Monit declares the
service dead and restarts it. What this means is that Director Health
Monitors that are on underprovisioned VMs and/or that have large numbers
of Deployments and/or Instances make take longer than 10 seconds to come
up, and may find themselves in an unending restart loop.

The guard around the '/unresponsive_agents' endpoint is added to
preserve the previous behavior of "Calls to /unresponsive_agents do not
succeed until the initial query of Director-managed deployments
completes.".

[#187938284] [JIRA] BOSH-296 "Health monitor continuously exiting for foundation with large number of VMs"

Signed-off-by: Aram Price <aram.price@broadcom.com>