Starts Health Monitor server after NATS is up
This commit moves the Health Monitor HTTP server start down to after
we've connected to NATS and subscribed to NATS messages.

Why? If for some reason NATS is nonfunctional, it'd be good for monit to
have never believed Health Monitor was healthy. And if it takes anywhere
near ten seconds to connect to a co-located NATS server and subscribe to
a handful of messages, something's dreadfully wrong with your system,
and you __should__ be seeing failures to try to alert you to this fact.

We do still start the HTTP server before we do a scan of all
deployments.

[#187938284] [JIRA] BOSH-296 "Health monitor continuously exiting for foundation with large number of VMs"