`#fake_app` method returns a stricter instance double
Previously a generic `double()` was being returned `as_null_object`
which allowed any number of method chains to be called. This commit
changes the object to be a set of instance doubles with one specific
method `#can_sign_urls?` explicitly stubbed on
`Bosh::Blobstore::BaseClient` as `false` which would have been the
value returned in the "null object" case.