Upgrade `sequel` from `5.29` => `5.85`
Because sequel now handles migrations ~better~ differently with sqlite
some older migrations had to be modified to accommodate the stricter
behavior/expectations of sequel + sqlite. As a consquence in the
migrations which were not functioning in sqlite we resorted to a backup,
recreate, repopulate strategy since `alter_table` and friends were not
behaving as expected with sqlite.

Sqlite was added to the Github Actions test matrix for the bosh-director
internal gem.

Signed-off-by: Brian Upton <brian.upton@broadcom.com>