CI: rework BufferedLogger into LoggingHelper
The primary change here is to remove the spec-global exception stubbing
of `Logger.new` and `Logging.logger`, presumably to prevent unexpected
loggers in the code base.

This spec-global exception stub is incompatible with the latest release
of Sinatra because one of its dependencies expects to be able to
instantiate `Logger`.

This commit also renames `logger` => `per_spec_logger` for clarity since
there are many instances of spec-local `logger` variables.