Remove special short timeout used for the agent_client during cck
It's been there since 2011 but we had reports that jobs will sometimes fail when
performing a CCK that succeed just fine during a normal deploy. We traced this down
to CCK having a shorter timeout and fewer retries but can't find any good
justification for this difference in agent_client behavior.

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>