Moves db data dirs off of tmpfs
We noticed that the Bosh Agent bosh-integration-tests Concourse Job was
failing. Postgres was reporting that it has run out of disk space.
Investigation revealed that both the Postgres and Mysql servers that the
tests set up are configured to put their data directories on tmpfs.

Looking through all available history reveals no solid justification for
this choice. Thinking through it now, we can come up with no good reason
to keep things this way.

The only alternative would be to *increase* the ramdisk size used by
Postgres, which would eat more of an already-scarce resource. Better to
just put the data directory back on disk.

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>